# Improves performance of prompt filtering