--- a/config/cmder.lua
+++ b/config/cmder.lua
@@ -136,7 +136,7 @@ end
  -- @return {bool}
 ---
 function get_git_status()
-    return os.execute("git diff --quiet --ignore-submodules HEAD")
+    return os.execute("git diff --quiet --ignore-submodules HEAD 2>nul")
 end
 
 function git_prompt_filter()