This PR dramatically improves performance of promp filtering for git and mercurial by using custom logic to pre-check if process calls is necessary. The overal `clink.prompt_filter` execution time is decreased from ~600 ms to:
- ~10 ms in non-repo directory
- ~130 ms for git repos
- ~350 ms for hg repos